+++ app/src/main/java/com/order/mall/data/network/shop/ShopOrder.java
@@ -22,14 +22,14 @@ public class ShopOrder {
 
     private String address;
     private String deliveryTime;
-    private int freight;
+    private float freight;
     private String id;
     private String mobile;
     private String name;
     private int num;
     private String payTime;
     private String picture;
-    private int price;
+    private float price;
     private String reciver;
     private int status;
     private int userId;
@@ -51,11 +51,11 @@ public class ShopOrder {
         this.deliveryTime = deliveryTime;
     }
 
-    public int getFreight() {
+    public float getFreight() {
         return freight;
     }
 
-    public void setFreight(int freight) {
+    public void setFreight(float freight) {
         this.freight = freight;
     }
 
@@ -107,11 +107,11 @@ public class ShopOrder {
         this.picture = picture;
     }
 
-    public int getPrice() {
+    public float getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(float price) {
         this.price = price;
     }
